metadata: fix SuperLoader Hdf5Loader issue
